requirements.txt
Django>=5.0,<6.0
gunicorn>=21.2
whitenoise>=6.6
dj-database-url>=2.2
python-dotenv>=1.0
psycopg2-binary>=2.9
django-ckeditor>=6.7
celery>=5.3
redis>=5.0
openai>=1.40.0
django-redis>=5.4.0
django-celery-results>=2.5.0
